browser/ss-facet-tags: pass tags and predicate to All Tags dialog (+25 squashed commits) Squashed commits: [7960ee0] browser/ss-facet-tags: tags styles [2af9b4f] browser/ss-facet-tags: remove stubData service
Stub data now defined in explore controller.
[a6c817a] browser/ss-facet-tags: start All Tags dialog
[97acbbb] browser/ss-facet-tags: remove menu template, fix repeats
[d89ffbb] browser/ss-facet-tags: fix styling for tags directive
[3278a78] browser/ss-facet-tags: add secondary sorting for tags
[cb60160] browser/ssFacetTags: tags as object, view has clear button
[3026971] browser/ss-facet-tags: move tag methods into directive link func
[4e76edb] browser/ss-facet-tags: typeahead filter by $viewValue
[0a109eb] move tags content to directive
[81e185e] add stub data to explore controller
[93a308f] add typeahead, selection to tags
[f9af4c0] browser/search: initial commit
[a506b72] browser/id-authenticate: restore mlAuth and include mlUserModel
[c06b5b6] temp hide mlAuth
[8d2fc9c] in-progress id-auth stuff
[3b7e2ab] browser: login dialog and account dropdown

Add dialog to present user with login prompt.

Add dropdown to display account information for logged-in user.
[6b2dc33] browser/auth: initial commit
[a76b6b8] browser/data: remove searchParams
[fa8586a] browser/http: re-introduce mlWaiter

mlWaiter helps track status of http requests and takes a bunch of
responsibility away from controllers to handle http-based promises
and statii.

mlWaiter was written prior to Angular 2.0 design docs becoming
available. As the direction for the features of mlWaiter has now been
sketched for the upcoming core, we will revisit mlWaiter in ea3 to
determine whether to adapt it to a pattern that follows where Angular
is heading. They have a different way of providing this feature.
[aec559f] browser/mlHttpAdapter: override method when POST /v1/login

Use http interceptor to override Angular http request to use
form-encoding when posting to /v1/login.
[5c86777] browser/data: initial commit
[5a6ed1c] fix: bootstrap font icons path
[26150b8] browser: update angular, angular mocks
[4c5aa9f] browser: only run the app from index.html

Running the app in unit tests causes it to attempt to load template
files and do other things we don't wish to do at test start.

Move the running of the app module back into index.html so that it is
not run in unit tests.