browser/http: re-introduce mlWaiter
mlWaiter helps track status of http requests and takes a bunch of
responsibility away from controllers to handle http-based promises
and statii.

mlWaiter was written prior to Angular 2.0 design docs becoming
available. As the direction for the features of mlWaiter has now been
sketched for the upcoming core, we will revisit mlWaiter in ea3 to
determine whether to adapt it to a pattern that follows where Angular
is heading. They have a different way of providing this feature.