browser/application: general clean up, first docs
Eliminate appSettings constant, assembledApp and configuredApp.

Replace with application module.

Configure settings that are driven by buildParams in the application
module as an angular config step.

Move CSRF enabled/disabled setting to mlHttpInterceptor as configured
in application module.

Move html5Mode setting to appRoutingProvider as configured in
application module.  Simplify code that sets it in appRoutingProvider.

Remove extraneous or unused appRouting functions.

Eliminate mlHttp module.

Rename marklogic.sample to _marklogic so that documentation makes sense.
TBD to fix documentation generation so that we are not forced down
this path.

Hardcode module names for simplicity so that lodash template code is not
found throughout the code.

Add basic documentation for:

* _marklogic/module
* _marklogic/mlHttpInterceptor
* app/module